Add apple-privacy-manifest.md for glTFast
This commit adds the apple-privacy-manifest file for the glTFast SDK per legal
requirements. If the SDK itself doesn't collect data, we don't need to document
the actual privacy manifest file. In this case we just clarify that the SDK does
not collect data per the Legal disclaimer, and list any additional SDKs that are
dependencies of this SDK.